chore: update to astro 5
