+++ mdx-plugins/index.ts
@@ -1,2 +1,1 @@
-export { rehypeCopy } from "./rehype-copy";
 export { remarkAsides } from "./remark-asides";
