--- mdx-plugins/rehype-copy.ts
@@ -1,15 +1,0 @@
-import type { Root } from "hast";
-import type { Plugin } from "unified";
-import { visit } from "unist-util-visit";
-
-export const rehypeCopy: Plugin<void[], Root> = () => {
-  return (tree) => {
-    visit(tree, "element", (node) => {
-      if (node.tagName !== "pre") {
-        return;
-      }
-
-      node.properties.copy = true;
-    });
-  };
-};
