+++ package.json
@@ -2,7 +2,7 @@
   "name": "luxass.dev",
   "type": "module",
   "private": true,
-  "packageManager": "pnpm@9.12.3",
+  "packageManager": "pnpm@9.15.0",
   "scripts": {
     "dev": "astro dev",
     "start": "astro dev",
@@ -14,50 +14,42 @@
   },
   "dependencies": {
     "@astrojs/check": "^0.9.4",
-    "@astrojs/mdx": "^3.1.9",
-    "@astrojs/rss": "^4.0.9",
-    "@astrojs/sitemap": "^3.2.1",
-    "@astrojs/solid-js": "^4.4.2",
-    "@vercel/analytics": "^1.3.2",
-    "astro": "^4.16.8",
-    "astro-icon": "^1.1.1",
-    "markdown-it": "^14.1.0",
+    "@vercel/analytics": "^1.4.1",
+    "astro": "^5.0.3",
+    "astro-icon": "^1.1.4",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-external-links": "^3.0.0",
     "rehype-slug": "^6.0.0",
-    "sanitize-html": "^2.13.1",
     "solid-js": "^1.9.3",
     "zod": "^3.23.8",
     "zod-error-utils": "^0.2.0"
   },
   "devDependencies": {
-    "@astrojs/vercel": "^7.8.2",
+    "@astrojs/vercel": "8.0.0",
     "@iconify-json/logos": "^1.2.3",
-    "@iconify-json/lucide": "^1.2.11",
+    "@iconify-json/lucide": "^1.2.18",
     "@iconify-json/mdi": "^1.2.1",
-    "@iconify-json/tabler": "^1.2.7",
-    "@luxass/eslint-config": "^4.12.1",
+    "@iconify-json/tabler": "^1.2.10",
+    "@luxass/eslint-config": "^4.14.0",
     "@types/hast": "^3.0.4",
-    "@types/markdown-it": "^14.1.2",
     "@types/mdast": "^4.0.4",
-    "@types/node": "^20.14.10",
-    "@types/sanitize-html": "^2.13.0",
-    "@unocss/eslint-plugin": "^0.63.6",
-    "@unocss/reset": "^0.63.6",
-    "astro-eslint-parser": "^1.0.3",
-    "eslint": "^9.14.0",
-    "eslint-plugin-astro": "^1.3.0",
-    "eslint-plugin-format": "^0.1.2",
+    "@types/node": "^20.17.9",
+    "@unocss/eslint-plugin": "^0.65.1",
+    "@unocss/reset": "^0.65.1",
+    "astro-eslint-parser": "^1.1.0",
+    "eslint": "^9.16.0",
+    "eslint-plugin-astro": "^1.3.1",
+    "eslint-plugin-format": "^0.1.3",
     "fontaine": "^0.5.0",
     "hastscript": "^9.0.0",
     "mdast-util-directive": "^3.0.0",
     "prettier-plugin-astro": "^0.14.1",
     "remark": "^15.0.1",
     "remark-directive": "^3.0.0",
-    "typescript": "^5.6.3",
+    "typescript": "^5.7.2",
     "unified": "^11.0.5",
     "unist-util-remove": "^4.0.0",
     "unist-util-visit": "^5.0.0",
-    "unocss": "^0.63.6"
+    "unocss": "^0.65.1"
   }
 }
