--- src/components/content/ProjectCard.astro
@@ -1,31 +1,0 @@
----
-import type { CollectionEntry } from "astro:content";
-import { Icon } from "astro-icon/components";
-
-export interface Props {
-  project: CollectionEntry<"projects">;
-}
-
-const {
-  project: {
-    data: { name, githubUrl, handle },
-  },
-} = Astro.props;
----
-
-<div
-  class="flex justify-between items-center border border-neutral-300 bg-neutral-100 dark:(border-neutral-700 bg-neutral-800) rounded p-3"
->
-  <a href={`/projects/${handle}`}>
-    <span>{name}</span>
-  </a>
-
-  <span class="flex items-center gap-3">
-    <a target="_blank" rel="noopener noreferrer" href={githubUrl}>
-      <Icon name="mdi:github" class="size-6" aria-label={`GitHub URL for ${name}`} />
-    </a>
-    <a href={`projects/${name}`}>
-      <Icon name="mdi:arrow-top-right-thin" class="size-6" aria-label={`Go to ${name} project page`} />
-    </a>
-  </span>
-</div>
