--- src/components/content/ProjectPreviewCard.astro
@@ -1,29 +1,0 @@
----
-import type { CollectionEntry } from "astro:content";
-
-export interface Props {
-  project: CollectionEntry<"projects">["data"];
-}
-
-const { project } = Astro.props;
-
-const { name, description, githubUrl } = project;
-const isContributor = false;
-
-let url = githubUrl;
-
-if (name != null) {
-  url = `/projects/${name}`;
-}
----
-
-<a
-  href={url}
-  class="flex flex-col border border-neutral-300 bg-neutral-100 dark:(border-neutral-700 bg-neutral-800) rounded p-3 gap-3"
->
-  <span class="text-(lg imp)">
-    {name}
-    {isContributor && <span class="text-sm"> (contributor)</span>}
-  </span>
-  <span>{description}</span>
-</a>
