--- src/env.d.ts
@@ -1,3 +1,0 @@
-/// <reference path="../.astro/types.d.ts" />
-/// <reference types="astro/client" />
-/// <reference types="../.astro/typed-api.d.ts" />
