+++ src/layouts/PageLayout.astro
@@ -1,23 +1,17 @@
 ---
-import "../styles/globals.css";
-import { ViewTransitions, fade } from "astro:transitions";
-import Footer from "../components/layout/Footer.astro";
-import Header from "../components/layout/Header.astro";
-
 export interface Props {
   title: string;
   description: string;
   image?: string;
   keywords?: string[];
-  class?: string;
 }
 
 const {
-  description = "personal website and blog of lucas nørgård",
-  keywords = [],
+  keywords,
+  description,
   image = new URL(`https://image.luxass.dev/api/image/text?width=300&height=300`),
-  class: clazz,
 } = Astro.props;
+
 let { title } = Astro.props;
 
 if (title.includes("luxass.dev") && import.meta.env.SITE_HOST === "luxass.com") {
@@ -28,13 +22,11 @@ if (title.includes("luxass.dev") && import.meta.env.SITE_HOST === "luxass.com") 
 <!doctype html>
 <html lang="en">
   <head>
-    <ViewTransitions transition:animate={fade({ duration: "0.2s" })} />
-
     <!-- global metadata -->
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width,initial-scale=1" />
-    <link rel="icon" href="/favicon.ico" />
     <meta name="generator" content={Astro.generator} />
+    <link rel="icon" href="/favicon.ico" />
 
     {import.meta.env.SITE_HOST === "luxass.com" && <link rel="canonical" href="https://luxass.dev" />}
 
@@ -43,8 +35,9 @@ if (title.includes("luxass.dev") && import.meta.env.SITE_HOST === "luxass.com") 
     <!-- primary meta tags -->
     <title>{title}</title>
     <meta name="title" content={title} />
-    <meta name="keywords" content={keywords.join(", ")} />
     <meta name="description" content={description} />
+
+    {keywords != null && <meta name="keywords" content={keywords.join(",")} />}
 
     <!-- open graph -->
     <meta property="og:type" content="website" />
@@ -66,11 +59,31 @@ if (title.includes("luxass.dev") && import.meta.env.SITE_HOST === "luxass.com") 
 
     <link rel="preload" href="/fonts/dm-sans.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
   </head>
-  <body>
-    <Header transition:animate="none" />
-    <main class="flex-1 p-4 prose" class:list={clazz} transition:animate="slide">
+  <body class="mx-auto max-w-3xl px-4">
+    <main class="pb-16 pt-8 leading-relaxed sm:pt-12">
       <slot />
     </main>
-    <Footer />
+    <footer class="flex place-content-between py-12">
+      <p class="text-zinc-450 text-sm">Copyright © 2024 Lucas Nørgård. All rights reserved.</p>
+      <a href="https://github.com/luxass/luxass.dev" class="text-main underline text-sm hover:text-zinc-450"
+        >Source code</a
+      >
+    </footer>
   </body>
+  <style>
+    html {
+      font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif,
+        "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+    }
+    code {
+      font-family:
+        ui-monospace,
+        SFMono-Regular,
+        SF Mono,
+        Menlo,
+        Consolas,
+        Liberation Mono,
+        monospace;
+    }
+  </style>
 </html>
