--- src/pages/404.astro
@@ -1,11 +1,0 @@
----
-import PageLayout from "../layouts/PageLayout.astro";
-
-export const prerender = true;
----
-
-<PageLayout title="404" description="404 - Page not found" class:list={["flex flex-col items-center justify-center"]}>
-  <h1 class="!text-8xl !text-blue-500">404</h1>
-  <p class="!mb-8 !text-2xl !mt-0">Page not found</p>
-  <a href="/">Go back</a>
-</PageLayout>
