+++ src/pages/index.astro
@@ -2,20 +2,10 @@
 import { Icon } from "astro-icon/components";
 import { getCollection } from "astro:content";
 import PostCard from "../components/content/PostCard.astro";
-import ProjectCard from "../components/content/ProjectCard.astro";
 import PageLayout from "../layouts/PageLayout.astro";
 import { sortPosts } from "../lib/utils";
 
 const posts = sortPosts(await getCollection("posts", ({ data }) => !data.draft)).slice(0, 4);
-const projects = await getCollection("projects").then((projects) => {
-  for (let i = projects.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [projects[i], projects[j]] = [projects[j], projects[i]];
-  }
-
-  return projects.slice(0, 6);
-});
-
 export const prerender = true;
 ---
 
@@ -48,7 +38,7 @@ export const prerender = true;
     <h2>Projects</h2>
 
     <div class="flex flex-col gap-4">
-      {projects.map((project) => <ProjectCard {project} />)}
+      <!-- {projects.map((project) => <ProjectCard {project} />)} -->
 
       <a href="/projects" class="flex items-center gap-2">
         All projects
