+++ src/pages/projects/index.astro
@@ -1,9 +1,5 @@
 ---
-import { getCollection } from "astro:content";
-import ProjectPreviewCard from "../../components/content/ProjectPreviewCard.astro";
 import PageLayout from "../../layouts/PageLayout.astro";
-
-const allProjects = await getCollection("projects");
 
 export const prerender = true;
 ---
@@ -12,7 +8,7 @@ export const prerender = true;
   <article class="prose">
     <h1>Projects</h1>
     <p>This is some of my projects.</p>
-    {
+    <!-- {
       allProjects.length > 0 ? (
         <div class="grid auto-cols-max grid-cols-1 mt-4 gap-4 sm:grid-cols-2 sm:gap-3" id="search-results">
           {allProjects.map(({ data: project }) => (
@@ -27,6 +23,6 @@ export const prerender = true;
           </a>
         </p>
       )
-    }
+    } -->
   </article>
 </PageLayout>
